테스트를 옮긴다
 - Symbol 테스트 클래스가 없어서 Symbol에 정의된 메서드의 테스트가
   있는건지 한번에 알기 힘들다. 이 테스트가 ChordParserTest에 있다.
 - ChordParser는 Symbol 메서드를 단순히 호출만 하고 있으므로 Symbol에
   정의된 메서드를 테스트하는 역할이 알맞지 않다고 판단했다.
 - 그래서 기존에 ChordParserTest에 정의된 테스트를 SymbolTest로 옮기고,
   ChordParserTest에는 Symbol에 정의한 메서드 호출을 확인하는 목적의
   테스트 하나만 남겼다.
